refactor: extract logs terminal ui in its own component
