+++ src/app/components/LogsTerminal/LogsTerminal.module.scss
@@ -1,10 +1,9 @@
 .logTerminal {
   height: 300px;
   overflow-y: scroll !important;
-  scrollbar-color: rgb(94, 107, 133) rgb(29, 39, 59);
+  scrollbar-color: rgb(102, 115, 130), rgb(29, 39, 59);
 }
 
 .wrapLines {
   text-wrap: wrap;
 }
-
